requirements.txt
python-dotenv
requests
youtube-transcript-api
openai
